Working on splitting out views.py into multiple separate files
Just tested on local dev and it's working so far
