+++ website/views/__init__.py
@@ -1,0 +1,12 @@
+from .faq import *
+from .index import *
+from .member import *
+from .news_listing import *
+from .news import *
+from .people import *
+from .project import *
+from .projects import *
+from .publications import *
+from .talks import *
+from .videos import *
+
